Added Toggle in UI to set the warp feature in the  playground
