+++ Assets/_Main/Scripts/Interfaces/IPlaygroundManager.cs
@@ -1,0 +1,7 @@
+﻿namespace _Main.Scripts.Interfaces
+{
+    public interface IPlaygroundManager
+    {
+        void SetWallsInteraction(bool p_b);
+    }
+}
